Propose a vision for the foundation
# Vision Statement

_In contrast to a mission statement which describes an organization's goals, a vision statement describes the future in which the organization accomplishes those goals. What would the world look like if the .NET Foundation is successful in its mission?_

A diverse, healthy, and active OSS community where project maintainers are well supported and contributors feel welcome. An ecosystem where .NET OSS is adopted in the enterprise, education, and personal projects. An ecosystem where the Foundation, its members, and the larger .NET OSS ecosystem work together to identify challenges to the mission and then collaborate on solutions.

A community where those that benefit the most from .NET OSS contribute back, whether it be through resources, time, or money. In this community, it is easy for anyone who wishes to contribute to do so in whatever way they can. It should be easy for companies to contribute financially to OSS and easy for project maintainers to receive that support.